新闻详情添加创建日期

+++ database/migrations/2018_07_25_172547_create_balance_records_table.php
@@ -16,8 +16,8 @@ class CreateBalanceRecordsTable extends Migration
         Schema::create('balance_records', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('user_id');
-            $table->integer('number');
-            $table->integer('surplus');
+            $table->decimal('number');
+            $table->decimal('surplus');
             $table->string('describe');
             $table->tinyInteger('type')->default(1)->comment('1收入 2支出');
             $table->timestamps();
